extract product reviews mock to a separate file
Similar to the product list mock, this extracts the ProductReview mock
from `products.ts` into `reviews.ts`. This should allow us to add more
mocks for the product reviews in the future and while retaining the
context of what the mocks in the file are for.
