+++ frontend/tests/jest/utils.tsx
@@ -67,30 +67,3 @@ export const renderWithAppContext = (
    ui: any,
    options?: RenderOptions
 ): RenderResult => render(ui, { wrapper: GlobalContextProvider, ...options });
-
-export const getReviewsResponse = (
-   type: 'filled' | 'empty' = 'filled'
-): ProductReviewData => {
-   let reviews: ProductReview[] = mockedReviews;
-
-   if (type === 'empty') {
-      reviews = [];
-   }
-
-   return {
-      reviews: reviews,
-      count: reviews.length,
-      average:
-         reviews.length == 0
-            ? 0
-            : reviews.reduce((total, next) => total + (next.rating || 0), 0) /
-              reviews.length,
-      groupedReviews: {
-         1: reviews.filter((review) => review.rating == 1).length,
-         2: reviews.filter((review) => review.rating == 2).length,
-         3: reviews.filter((review) => review.rating == 3).length,
-         4: reviews.filter((review) => review.rating == 4).length,
-         5: reviews.filter((review) => review.rating == 5).length,
-      },
-   };
-};
